backend/requirements.txt
# ================================
# FG Aesthetic NFC Loyalty System
# Backend Dependencies
# ================================

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Environment & Config
python-dotenv==1.0.0

# Database (Supabase)
supabase==2.0.2

# Data Validation
pydantic==2.5.3
email-validator==2.1.0
